moved stories to stories folder and fixed react missing keys
