--- caption/src/index.stories.tsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import { Caption } from './index';
-
-export default {
-  parameters: {
-    layout: "centered",
-  },
-};
-
-export const caption =  () => <Caption text="test caption" />;
