+++ caption/stories/index.stories.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import { Caption } from '../index';
+
+export default {
+  parameters: {
+    layout: "centered",
+  },
+};
+
+export const caption = () => <Caption text="test caption" />;
