+++ showcases/doc/showcases.mdx
@@ -3,7 +3,7 @@ import { mdx, MDXCreateElement } from '@mdx-js/react';
 import { Props, Description } from '~/props/dist/index';
 import { Showcases } from '../dist';
 import { CssLayout } from '~/mdx-layout-css';
-import '../dist/showcases-stories.css';
+import '../stories/showcases-stories.css';
 
 export default CssLayout;
