+++ text/stories/index.stories.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Text } from './index';
+import { Text } from '../index';
 import './text-stories.css';
